refactor(voice): extract lib/voice module and add masked API key support
- Create lib/voice/interfaces.ts, svc.ts, utils.ts from inline code
- Move API functions from lib/admin/voice/svc.ts to lib/voice/svc.ts
- Add masked api_key field to VoiceProviderView (backend + frontend)
- Use masked key as initial form value for editing (api_key_changed
  detection by comparing against initial value)
- Clean up shared.tsx imports to use lib/voice/ modules
