+++ web/src/hooks/useVoiceProviders.ts
@@ -1,19 +1,9 @@
 import useSWR from "swr";
 import { errorHandlingFetcher } from "@/lib/fetcher";
 import { SWR_KEYS } from "@/lib/swr-keys";
+import type { VoiceProviderView } from "@/lib/voice/interfaces";
 
-export interface VoiceProviderView {
-  id: number;
-  name: string;
-  provider_type: string;
-  is_default_stt: boolean;
-  is_default_tts: boolean;
-  stt_model: string | null;
-  tts_model: string | null;
-  default_voice: string | null;
-  has_api_key: boolean;
-  target_uri: string | null;
-}
+export type { VoiceProviderView };
 
 export function useVoiceProviders() {
   const { data, error, isLoading, mutate } = useSWR<VoiceProviderView[]>(
